Action Filter
This visualization presents the performance of ‘Action Filters’. Using this filter, I have visualized ‘Product Category by Sales’ and ‘Product Sub-Category Sales’. When managers click on any category, they will see the detailed sales amounts of the sub-categories under that category.
In the graph on the right side, the X-axis denotes the supplies by category, and the Y-axis denotes the sales amount of the superstore. After selecting any category, it will display the sub-categories on the X-axis and their corresponding sales amounts on the Y-axis. The graph on the left side shows the outcomes of the action filter.